[Refactor] Extract eval_conjuncts from Exec
