+++ be/src/exec/hdfs_scanner/hdfs_scanner_json.cpp
@@ -15,6 +15,7 @@
 #include "exec/hdfs_scanner/hdfs_scanner_json.h"
 
 #include "common/simdjson_util.h"
+#include "exprs/chunk_predicate_evaluator.h"
 #include "formats/avro/nullable_column.h"
 #include "formats/json/json_utils.h"
 #include "formats/json/nullable_column.h"
@@ -222,7 +223,7 @@ Status HdfsJsonScanner::do_get_next(RuntimeState* runtime_state, ChunkPtr* chunk
 
         for (auto& [_, ctxs] : _scanner_ctx.conjunct_ctxs_by_slot) {
             SCOPED_RAW_TIMER(&_app_stats.expr_filter_ns);
-            RETURN_IF_ERROR(ExecNode::eval_conjuncts(ctxs, chunk->get()));
+            RETURN_IF_ERROR(ChunkPredicateEvaluator::eval_conjuncts(ctxs, chunk->get()));
             if ((*chunk)->num_rows() == 0) {
                 break;
             }
