+++ be/src/exec/pipeline/exchange/split_local_exchange.cpp
@@ -16,7 +16,7 @@
 
 #include <memory>
 
-#include "exec/exec_node.h"
+#include "exprs/chunk_predicate_evaluator.h"
 #include "exprs/expr_executor.h"
 
 namespace starrocks::pipeline {
@@ -54,8 +54,8 @@ Status SplitLocalExchanger::push_chunk(const ChunkPtr& chunk, int32_t sink_drive
     for (size_t i = _split_expr_ctxs.size() - 1; i > 0; i--) {
         Filter chunk_filter(cur_chunk->num_rows(), 1);
         auto& expr_ctx = _split_expr_ctxs[i];
-        ASSIGN_OR_RETURN(size_t new_chunk_size,
-                         ExecNode::eval_conjuncts_into_filter({expr_ctx}, cur_chunk.get(), &chunk_filter));
+        ASSIGN_OR_RETURN(size_t new_chunk_size, ChunkPredicateEvaluator::eval_conjuncts_into_filter(
+                                                        {expr_ctx}, cur_chunk.get(), &chunk_filter));
         // all false for expr_ctx
         if (new_chunk_size == 0) {
             continue;
