+++ be/src/exec/pipeline/scan/schema_chunk_source.cpp
@@ -19,6 +19,7 @@
 
 #include "exec/schema_scanner.h"
 #include "exec/workgroup/work_group.h"
+#include "exprs/chunk_predicate_evaluator.h"
 #include "runtime/descriptors_ext.h"
 
 namespace starrocks::pipeline {
@@ -197,7 +198,7 @@ Status SchemaChunkSource::_read_chunk(RuntimeState* state, ChunkPtr* chunk) {
             SCOPED_TIMER(_filter_timer);
             auto& conjunct_ctxs = _ctx->conjunct_ctxs();
             if (!conjunct_ctxs.empty()) {
-                RETURN_IF_ERROR(ExecNode::eval_conjuncts(conjunct_ctxs, chunk_dst.get()));
+                RETURN_IF_ERROR(ChunkPredicateEvaluator::eval_conjuncts(conjunct_ctxs, chunk_dst.get()));
             }
         }
         row_num = chunk_dst->num_rows();
